// src/DockiUp.Frontend/src/app/api/model/deploymentDto.ts
/**
 * DockiUp.API | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DeploymentTrigger } from './deploymentTrigger';
import { DeploymentStatus } from './deploymentStatus';


export interface DeploymentDto { 
    id: string;
    projectId: string;
    trigger: DeploymentTrigger;
    status: DeploymentStatus;
    actorName: string | null;
    commitBefore: string | null;
    commitAfter: string | null;
    commitMessage: string | null;
    targetCommit: string | null;
    createdAt: string;
    startedAt: string | null;
    finishedAt: string | null;
    error: string | null;
    log?: string | null;
}
export namespace DeploymentDto {
}